fix: delete styled file
