+++ src/components/common/Button/Button.tsx
@@ -1,5 +1,10 @@
-import * as S from "./Button.styled";
-import type { ButtonProps } from "./Button.types";
+import styled from "@emotion/styled";
+import { css } from "@emotion/react";
+
+import colors from "styles/color";
+import fonts from "styles/font";
+
+import { ButtonProps, ButtonRounded } from "./Button.types";
 
 export default function Button({
   textColor = "white",
@@ -10,7 +15,7 @@ export default function Button({
   ...restProps
 }: ButtonProps) {
   return (
-    <S.Button
+    <_Button
       textColor={textColor}
       backgroundColor={backgroundColor}
       fontSize={fontSize}
@@ -18,6 +23,37 @@ export default function Button({
       {...restProps}
     >
       {children}
-    </S.Button>
+    </_Button>
   );
 }
+
+const _Button = styled.button<ButtonProps>`
+  padding: 12px 24px;
+
+  ${({ textColor = "white" }) => css`
+    color: ${colors[textColor]};
+  `};
+
+  ${({ backgroundColor = "primary" }) => css`
+    background-color: ${colors[backgroundColor]};
+  `};
+
+  ${({ fontSize = "16_600" }) => css`
+    ${fonts[fontSize]}
+  `};
+
+  ${({ rounded = "small" }) => css`
+    border-radius: ${BUTTON_ROUNDED_MAP[rounded]};
+  `};
+
+  &:disabled {
+    color: ${colors.white};
+    background-color: ${colors.textMuted};
+  }
+`;
+
+const BUTTON_ROUNDED_MAP: Record<ButtonRounded, string> = {
+  small: "12px",
+  medium: "24px",
+  large: "38px",
+};
