+++ src/components/common/Button/Button.types.ts
@@ -1,4 +1,5 @@
 import type { ComponentPropsWithoutRef } from "react";
+
 import colors from "styles/color";
 import fonts from "styles/font";
